# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Problem Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Problem Management API goal is to provide\
    \ the ability to manage problems in the Service domain. \n\n### Resource \n- serviceProblem\n\
    \n### Operations\nService Problem Management API performs the following operations\
    \ on the resource :\n- Retrieve an entity or a collection of entities depending\
    \ on filter criteria\n- Partial update of an entity (including updating rules)\n\
    - Create an entity (including default values and creation rules)\n- Delete an\
    \ entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Problem Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /serviceProblemEventRecord:
    get:
      tags:
      - "serviceProblemEventRecord"
      summary: "List or find ServiceProblemEventRecord objects"
      description: "This operation list or find ServiceProblemEventRecord entities"
      operationId: "listServiceProblemEventRecord"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceProblemEventRecord"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblemEventRecord"
  /serviceProblemEventRecord/{id}:
    get:
      tags:
      - "serviceProblemEventRecord"
      summary: "Retrieves a ServiceProblemEventRecord by ID"
      description: "This operation retrieves a ServiceProblemEventRecord entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveServiceProblemEventRecord"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceProblemEventRecord"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceProblemEventRecord"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceProblemEventRecord"
...
definitions:
  Any: {}
...
  ServiceProblemEventRecord:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Identifier of the service problem event record."
      href:
        type: "string"
        description: "reference to this resource"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time at which the event occurred"
      eventType:
        type: "string"
        description: "Type of the recorded event"
      recordTime:
        type: "string"
        format: "date-time"
        description: "Time at which the record was created"
      notification:
        description: "A notification from the possible notifications for Service Problem\
          \ (such as creation, status change, information required, change)"
        $ref: "#/definitions/Any"
      serviceProblem:
        description: "The service problem to which this record applies"
        $ref: "#/definitions/ServiceProblemRef"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "A record of an event (related to a service problemn) received from\
      \ another system"
    example:
      notification: {}
      serviceProblem:
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        correlationId: "correlationId"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      recordTime: "2000-01-23T04:56:07.000+00:00"
      '@baseType': "@baseType"
      '@type': "@type"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      id: "id"
      href: "href"
      eventType: "eventType"
      '@schemaLocation': "http://example.com/aeiou"
  ServiceProblemRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the Problem"
      href:
        type: "string"
        description: "Reference of the Problem"
      correlationId:
        type: "string"
        description: "Additional identifier coming from an external system"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      correlationId: "correlationId"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"